+++ src/components/App/VoucherifyButton.js
@@ -3,7 +3,13 @@ import Button from '@material-ui/core/Button';
 import Tooltip from '@material-ui/core/Tooltip';
 import PropTypes from 'prop-types';
 
-const VoucherifyButton = ({ code, text, onClickFunction }) => {
+const VoucherifyButton = ({
+	code,
+	text,
+	onClickFunction,
+	specialText,
+	tooltip,
+}) => {
 	const [currentText, setCurrentText] = useState(text || null);
 	const [tooltipTitle, setTitle] = useState('Click to copy');
 
@@ -18,7 +24,19 @@ const VoucherifyButton = ({ code, text, onClickFunction }) => {
 		setCurrentText(code);
 	};
 
-	if (code) {
+	if (specialText && code && !tooltip) {
+		return (
+			<Button
+				className="voucherifyButtonOrange"
+				onClick={() => {
+					navigator.clipboard.writeText(code);
+					handleTooltipTitle();
+				}}
+			>
+				{currentText}
+			</Button>
+		);
+	} else if (code) {
 		return (
 			<Tooltip title={tooltipTitle}>
 				<Button